# docs/api/openapi-dar.yaml
openapi: 3.1.0
info:
  title: Data Access Register
  description: "Core APIs for the Central Data Access Register.\n\nThe register records\
    \ all lawful access to customer energy data by Controllers,\nregardless of the\
    \ legal basis on which that access rests. It is not exclusively\na consent register\
    \ — legitimate interests, public task, legal obligation, and\ncontract are all\
    \ within scope alongside consent-based access.\n\n**Jurisdiction:** This is a\
    \ UK-only solution. All Controllers, customers, and\nmeter points are within Great\
    \ Britain. The governing data protection framework\nis **UK GDPR** and the **Data\
    \ Protection Act 2018**. EU GDPR does not apply.\nThe jurisdiction field is fixed\
    \ to `GB` and does not need to be supplied.\n\n## Purpose\n\nThe register provides\
    \ a single, auditable source of truth for:\n\n- **Controllers** — to record and\
    \ demonstrate the legal basis under which they\n  access customer meter data.\n\
    - **Data Providers** — to verify that a Controller has a registered, active access\n\
    \  record before releasing data.\n- **Customers** — to view all parties registered\
    \ as having lawful access to their\n  meter point, on request via any Controller.\n\
    \n## Party Model\n\n| Party | UK GDPR Role | Description |\n|---|---|---|\n| **Data\
    \ User** | Processor | Operates the register. Fetches and processes meter data\
...
    \ Current version: `v1`.\n"
  version: 0.0.13
...
servers:
...
paths:
  /identity-records:
    get:
      operationId: lookupIdentityRecords
      summary: Look Up Identity Records
      description: 'Returns Identity Records matching the supplied lookup parameters.
        Used by

        a Data User to locate an existing record for a returning customer before

        initiating re-identification.


        At least one of `mpxn` or `email` must be supplied. Supplying both narrows

        results to records where both values match.


        Results are scoped to the authenticated Data User — a Data User cannot

        look up records created by another Data User.


        **Note on email lookup:** The register stores a one-way hash of the email

        address for lookup purposes. The plaintext email is never returned in

        responses. A match confirms a record exists for that address; the Data User

        should already hold the email in their own system.

        '
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: mpxn
        in: query
        required: false
        description: Filter by meter point identifier.
        schema:
          $ref: '#/components/schemas/MPxN'
      - name: email
        in: query
        required: false
        description: 'Filter by customer email address. Matched against a stored hash
          — the plaintext is not retained in results.

          '
        schema:
          type: string
          format: email
      responses:
        '200':
          description: Matching identity records returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LookupIdentityRecordsResponse'
              example:
                response:
                  resource: /v1/identity-records?mpxn=1234567890123
                  timestamp: '2026-03-24T09:00:00Z'
                  transaction-id: tid_abc123def456abc123def456
                identity-records:
                - ir: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                  created-at: '2024-01-15T09:30:00Z'
                  pii-principal:
                    mpxn: '1234567890123'
                    move-in-date: '2022-06-30'
                  expressed-by: data-subject
                  has-email: true
                  credentials:
                  - credential-id: cred_1a2b3c4d5e6f1a2b3c4d5e6f
                    registered-at: '2024-01-15T09:35:00Z'
        '400':
          description: Neither `mpxn` nor `email` was supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createIdentityRecord
      summary: Create an Identity Record
      description: "Creates an Identity Record holding the person-property relationship\
        \ for a\nconsent or access registration. Returns an `ir` key that must be\
        \ supplied\nas `record-metadata.identity-record-ref` when creating the linked\
        \ AccessRecord.\n\nThe Identity Record holds:\n- `pii-principal` — MPxN, move-in\
        \ date, and optionally address.\n- `expressed-by` — whether the data subject\
        \ or an authorised representative\n  initiated the registration.\n- `principal-verification`\
        \ — identity verification evidence (method, outcome,\n  reference). Strongly\
        \ recommended; null if no structured verification was performed.\n- `email`\
        \ — stored as a one-way hash for future magic-link re-identification\n  and\
        \ lookup. Optional but required if magic-link re-identification is to be used.\n\
        - `initiate-passkey-registration` — if `true`, the response includes a\n \
        \ `passkey-registration-redirect` containing a short-lived URL to\n  `id.central.consent`.\
        \ Redirect the customer to this URL — the register\n  completes the WebAuthn\
        \ ceremony on its own origin, stores the public key,\n  and redirects back\
        \ with `?dar-passkey-token={token-ref}`. Call\n  `GET /identity-records/{ir}/re-identify/{token-ref}`\
        \ once to confirm.\n  Supply `passkey-return-url` to control where the customer\
        \ lands after\n  the ceremony.\n\nIdentity Records are scoped to the authenticated\
        \ Data User and are never\nexposed to the unauthenticated `GET /access-records/{ak}`\
        \ endpoint used by\nData Providers.\n"
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdentityRecordRequest'
            examples:
              with-email-and-passkey:
                summary: With email and passkey registration initiated
                value:
                  pii-principal:
                    mpxn: '1234567890123'
                    move-in-date: '2022-06-30'
                    address:
                      addressLine1: 221B Baker Street
                      townCity: London
                      postcode: NW1 6XE
                  expressed-by: data-subject
                  email: customer@example.com
                  initiate-passkey-registration: true
                  principal-verification:
                    method: credit-card
                    verified-at: '2023-11-07T05:30:45Z'
                    outcome: verified
                    reference: ch_3ABC123xyz
                    submitted: XXXX-XXXX-XXXX-4242
                    verified-against: Stripe customer record cus_ABC123
                    detail:
                      last4: '4242'
                      brand: visa
                      stripe_customer: cus_ABC123
              email-only:
                summary: With email, no passkey
                value:
                  pii-principal:
                    mpxn: '1234567890123'
                    move-in-date: '2020-01-01'
                  expressed-by: data-subject
                  email: customer@example.com
                  principal-verification: null
              minimal:
                summary: No re-identification credentials
                value:
                  pii-principal:
                    mpxn: '1234567890123'
                    move-in-date: '2020-01-01'
                  expressed-by: data-subject
                  principal-verification: null
      responses:
        '201':
          description: 'Identity record created successfully. If `initiate-passkey-registration`

            was `true`, `passkey-registration-redirect` is populated — redirect the

            customer to `redirect-url`. The register completes the WebAuthn ceremony

            on `id.central.consent` and sends the customer back to `return-url` with

            `?dar-passkey-token={token-ref}`. Call

            `GET /identity-records/{ir}/re-identify/{token-ref}` once to confirm.

            '
          headers:
            Location:
              description: URL of the newly created identity record.
              schema:
                type: string
                example: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIdentityRecordResponse'
              examples:
                with-passkey-redirect:
                  summary: Response when passkey registration was initiated
                  value:
                    response:
                      resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3
                      timestamp: '2024-01-15T09:30:00Z'
                      transaction-id: tid_691df0c788ca043403b7fa90
                    ir: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                    passkey-registration-redirect:
                      redirect-url: https://id.central.consent/passkey/register?session=pks_a1b2c3d4e5f6a1b2c3d4e5f6
                      token-ref: mlr_9f8e7d6c5b4a9f8e7d6c5b4a
                      expires-at: '2024-01-15T09:35:00Z'
                without-passkey-redirect:
                  summary: Response when no passkey registration requested
                  value:
                    response:
                      resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3
                      timestamp: '2024-01-15T09:30:00Z'
                      transaction-id: tid_691df0c788ca043403b7fa90
                    ir: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                    passkey-registration-redirect: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /identity-records/{ir}:
    get:
      operationId: getIdentityRecord
      summary: Retrieve an Identity Record
      description: 'Returns the full Identity Record for a given `ir` key. Only accessible
        to

        the authenticated Data User that created the record.


        The `email` field is never returned in plaintext — `has-email` indicates

        whether an email address is registered on the record. The `credentials`

        array returns passkey metadata (credential ID, registered-at) but never

        the public key material itself.

        '
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: ir
        in: path
        description: The identity record key issued at creation.
        required: true
        schema:
          $ref: '#/components/schemas/IdentityRecordKey'
      responses:
        '200':
          description: Identity record found and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIdentityRecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteIdentityRecord
      summary: Anonymise an Identity Record
      description: 'Anonymises an Identity Record in response to a GDPR Art. 17 erasure

        request. The record is retained in a tombstoned state for audit continuity

        — the `ir` key remains valid and the linked AccessRecord continues to

        reference it — but all PII fields (`pii-principal`, `email`, `credentials`,

        and `principal-verification`) are permanently destroyed.


        **Pre-condition:** All AccessRecords linked to this Identity Record must be

        in `REVOKED` or `EXPIRED` state. Attempting to anonymise an Identity Record

        linked to an `ACTIVE` access record returns `409`.


        This operation is irreversible.

        '
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: ir
        in: path
        description: The identity record key to anonymise.
        required: true
        schema:
          $ref: '#/components/schemas/IdentityRecordKey'
      responses:
        '200':
          description: Identity record anonymised. All PII permanently destroyed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnonymiseIdentityRecordResponse'
              example:
                response:
                  resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3
                  timestamp: '2026-03-24T10:00:00Z'
                  transaction-id: tid_abc123def456abc123def456
                ir: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                anonymised-at: '2026-03-24T10:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'One or more linked AccessRecords are still `ACTIVE`. Revoke
            or allow all linked access records to expire before anonymising.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /identity-records/{ir}/credentials/{credentialId}:
    delete:
      operationId: removePasskeyCredential
      summary: Remove a Passkey Credential
      description: 'Removes a registered passkey credential from an Identity Record.
        The

        credential can no longer be used for passkey assertion after removal.


        If no credentials remain on the record after removal, passkey-based

        re-identification is no longer available for this Identity Record.

        Magic-link re-identification remains available if an email address is

        registered.

        '
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: ir
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/IdentityRecordKey'
      - name: credentialId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/CredentialId'
      responses:
        '204':
          description: Credential removed. No response body.
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /identity-records/{ir}/re-identify:
    post:
      operationId: reIdentify
      summary: Initiate Re-identification
      description: 'Initiates or completes a re-identification flow for a returning
        customer,

        reconnecting them to their existing Identity Record without re-collecting

        all their details.


        **Three methods are supported:**


        **`magic-link`** — requires `email` to be stored on the Identity Record.

        The register dispatches a short-lived signed URL to the stored email address.

        The link expires after 15 minutes and is single-use.


        Supply an optional `redirect-url` (pre-registered for the Data User''s DUID)

        to send the customer back to the Data User''s application after clicking —

        the register appends `?dar-reid-token={token-ref}` so the application can

        confirm re-identification with a single call rather than polling.


        Without `redirect-url`, the customer lands on a confirmation page at

        `central.consent` and the Data User polls

        `GET /identity-records/{ir}/re-identify/{token-ref}` for confirmation.


        **`passkey-assert`** — requires at least one credential registered on the

        Identity Record. The register returns a WebAuthn assertion challenge. The

        Data User passes this to `navigator.credentials.get()`, then submits the

        signed assertion response back to this endpoint with `method: passkey-assert-complete`

        to verify and confirm re-identification.


        **`passkey-register`** — used when the customer has no existing passkey

        on the record (e.g. new device). Returns a WebAuthn registration challenge.

        The Data User completes the ceremony via `POST /identity-records/{ir}/credentials`

        to store the new public key, then the re-identification is confirmed.


        Re-identification does not modify the core Identity Record fields. It is a

        verification step confirming the returning individual is the same principal

        as originally recorded.

        '
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: ir
        in: path
        description: The identity record key for the returning principal.
        required: true
        schema:
          $ref: '#/components/schemas/IdentityRecordKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReIdentifyRequest'
            examples:
              magic-link-with-redirect:
                summary: Magic-link with redirect back to Data User app
                value:
                  method: magic-link
                  redirect-url: https://app.bright-energy.com/renew/confirmed
              magic-link-portal:
                summary: Magic-link without redirect (customer lands on central.consent)
                value:
                  method: magic-link
              passkey-assert:
                summary: Initiate passkey assertion (existing credential)
                value:
                  method: passkey-assert
                  passkey-return-url: https://app.bright-energy.com/renew/confirmed
              passkey-register:
                summary: Initiate passkey registration (new device)
                value:
                  method: passkey-register
                  passkey-return-url: https://app.bright-energy.com/onboard/passkey-done
      responses:
        '200':
          description: "Re-identification flow initiated.\n\n- For `magic-link`: link\
            \ dispatched, `magic-link` object populated.\n- For `passkey-assert` or\
            \ `passkey-register`: `passkey` redirect object\n  populated — redirect\
            \ the customer to `passkey.redirect-url` immediately.\n"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReIdentifyResponse'
              examples:
                magic-link-dispatched:
                  summary: Magic link dispatched (with redirect-url)
                  value:
                    response:
                      resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3/re-identify
                      timestamp: '2026-03-24T09:00:00Z'
                      transaction-id: tid_abc123def456abc123def456
                    method: magic-link
                    magic-link:
                      dispatched-to: c*****r@example.com
                      expires-at: '2026-03-24T09:15:00Z'
                      token-ref: mlr_9f8e7d6c5b4a9f8e7d6c5b4a
                      redirect-url: https://app.bright-energy.com/renew/confirmed
                    passkey: null
                passkey-assert-redirect:
                  summary: Passkey assertion — redirect to id.central.consent
                  value:
                    response:
                      resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3/re-identify
                      timestamp: '2026-03-24T09:00:00Z'
                      transaction-id: tid_abc123def456abc123def456
                    method: passkey-assert
                    magic-link: null
                    passkey:
                      redirect-url: https://id.central.consent/passkey/assert?session=pks_a1b2c3d4e5f6a1b2c3d4e5f6
                      token-ref: mlr_1a2b3c4d5e6f1a2b3c4d5e6f
                      expires-at: '2026-03-24T09:05:00Z'
                      return-url: https://app.bright-energy.com/renew/confirmed
        '400':
          description: 'Invalid method, missing required fields, challenge expired
            or already used, or assertion response failed verification.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'The Identity Record has been anonymised and cannot be used
            for re-identification.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: '`magic-link` requested but no email address is stored on this
            Identity Record. `passkey-assert` requested but no credentials are registered.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /identity-records/{ir}/re-identify/{tokenRef}:
    get:
      operationId: pollReIdentifyStatus
      summary: Confirm Re-identification Status
      description: 'Confirms the status of a magic-link or passkey re-identification
        flow.

        The `token-ref` is returned in the `magic-link` or `passkey` object from

        `POST /identity-records/{ir}/re-identify`, and also appended as a query

        parameter when the register redirects the customer back to the Data User''s

        application.


        **With a `redirect-url` or `passkey-return-url`** — the register appends

        `?dar-reid-token={token-ref}` (magic link) or `?dar-passkey-token={token-ref}`

        (passkey) to the return URL. Read the token from the query string and call

        this endpoint **once** — no polling required.


        **Without a return URL** — poll this endpoint until `status` is `confirmed`

        or `expired`.


        **Suggested polling interval:** every 3–5 seconds. Stop polling on

        `confirmed`, `expired`, or after 20 minutes.

        '
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: ir
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/IdentityRecordKey'
      - name: tokenRef
        in: path
        required: true
        description: The `token-ref` returned by the magic-link initiation call.
        schema:
          type: string
          pattern: ^mlr_[0-9a-f]{24}$
          example: mlr_9f8e7d6c5b4a9f8e7d6c5b4a
      responses:
        '200':
          description: Status returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReIdentifyStatusResponse'
              examples:
                pending:
                  summary: Awaiting customer redemption
                  value:
                    response:
                      resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3/re-identify/mlr_9f8e7d6c5b4a9f8e7d6c5b4a
                      timestamp: '2026-03-24T09:05:00Z'
                      transaction-id: tid_aabbccddeeff001122334455
                    status: pending
                    confirmed-at: null
                confirmed:
                  summary: Customer redeemed the link
                  value:
                    response:
                      resource: /v1/identity-records/ir_a3c5e7f9b1d3a3c5e7f9b1d3/re-identify/mlr_9f8e7d6c5b4a9f8e7d6c5b4a
                      timestamp: '2026-03-24T09:07:30Z'
                      transaction-id: tid_112233445566778899aabbcc
                    status: confirmed
                    confirmed-at: '2026-03-24T09:07:28Z'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /identity-records/exists:
    get:
      operationId: checkIdentityRecordExists
      summary: Check Identity Record Existence by MPxN
      description: 'Returns a boolean indicating whether an Identity Record exists
        for the given MPxN across all Data Users. No `ir` key or PII is returned.


        Used by a Data User before initiating cross-DUID re-identification — if `exists`
        is `true`, the Data User can call `POST /identity-records/reidentify` to initiate
        a re-identification challenge without needing to know which Data User created
        the record.'
      tags:
      - Identity Records
      security:
      - bearerAuth: []
      parameters:
      - name: mpxn
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/MPxN'
      responses:
        '200':
          description: Existence check result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    $ref: '#/components/schemas/ResponseMetadata'
                  mpxn:
                    $ref: '#/components/schemas/MPxN'
                  exists:
                    type: boolean
                required:
                - response
                - mpxn
                - exists
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Base64-encode `account_id:secret_key`. Used only on `GET /auth/token`.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from `GET /auth/token`. Pass as `Authorization: Bearer <token>`.
        Expires after 7200s.'
  responses:
...
    Unauthorised:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    MPxN:
      type: string
      pattern: ^(?:[0-9A-HJ-NPR-Z]{2}[0-9]{8,10}|[0-9]{10})$
      description: Meter Point Administration Number (MPAN) or Meter Point Reference
        Number (MPRN).
      example: '1234567890123'
...
    LookupIdentityRecordsResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/ResponseMetadata'
        identity-records:
          type: array
          items:
            $ref: '#/components/schemas/IdentityRecord'
      required:
      - response
      - identity-records
...